Rename BaseDriver to PDODriver
Rename BaseDriver to the more implicit PDODriver. Remove Str class from
being able to be passed in. Str generally won’t change.
